Style minimum appliqué

+++ supprimer_plat.php
@@ -1,6 +1,13 @@
 <?php
 session_start();
 require_once "config/connexionBDD.php";
+
+if($_SESSION['droits'] != 2)
+{
+    header('Location: index.php');
+    exit();
+}
+
 $id_plat = trim(htmlspecialchars($_GET['id']));
 $requete = $connexion->prepare("DELETE FROM plats WHERE id_plat = :id_plat");
 $requete->execute(['id_plat' => $id_plat]);
